Moved to Google Location API
- Changed radius from double to float, to match Google’s API spec

+++ app/src/main/java/com/harrymt/productivitymapping/ZoneEditActivity.java
@@ -67,7 +67,7 @@ public class ZoneEditActivity extends FragmentActivity implements GoogleMap.OnMa
         LatLng pos = currentCircle.centerMarker.getPosition();
         zoneToEdit.lat = pos.latitude;
         zoneToEdit.lng = pos.longitude;
-        zoneToEdit.radiusInMeters = currentCircle.radius;
+        zoneToEdit.radiusInMeters = (float) currentCircle.radius;
 
         zoneIntent.putExtra("zone", zoneToEdit);
         startActivityForResult(zoneIntent, REQUEST_CODE_SET_ZONE_PREFS);
